requirements.txt
apache-airflow==2.10.0
dbt-core==1.8.7
dbt-bigquery==1.8.2
google-cloud-bigquery==3.30.0
pandas==2.2.3
streamlit==1.44.1
python-dotenv==1.0.1
